requirements2.txt
alembic==1.6.3
appdirs==1.4.4
attributedict==0.3.0
attrs==21.2.0
blessings==1.7
blinker==1.4
cachelib==0.9.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
codecov==2.1.11
colorama==0.4.4
coloredlogs==15.0
colour-runner==0.1.1
config2==0.3.2
configparser==5.0.2
coverage==5.5
crayons==0.4.0
cryptography==3.4.7
deepdiff==5.5.0
deepmerge==0.3.0
distlib==0.3.1
docopt==0.6.2
dominate==2.6.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.0.12
Flask==2.2.2
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-MySQL==1.5.2
flask-nav==0.6
flask-session2==1.3.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
humanfriendly==9.1
idna==2.10
importlib-metadata==6.0.0
iniconfig==1.1.1
inspecta==0.1.3
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.1.2
more-itertools==8.8.0
mybad==0.1.4
openpyxl==3.0.10
ordered-set==4.0.2
packaging==20.9
peewee==3.14.4
peewee-migrations==0.3.19
pluggy==0.13.1
py==1.10.0
py-solc-ast==1.2.9
pycparser==2.20
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==7.2.0
pytest-base-url==1.4.2
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-watch==4.2.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2022.7.1
PyYAML==5.4.1
requests==2.25.1
rootpath==0.1.1
selenium==3.141.0
setupextras==0.1.5
six==1.16.0
SQLAlchemy==1.4.15
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tox==3.23.1
Unidecode==1.3.6
urllib3==1.26.4
virtualenv==20.4.6
visitor==0.1.3
watchdog==2.1.2
wcwidth==0.2.5
webdriver-manager==3.4.2
Werkzeug==2.2.2
wtf-peewee==3.0.4
WTForms==2.3.3
XlsxWriter==3.0.3
zipp==3.11.0
